[Topgg] Finalize module framework
