+++ bot/modules/topgg/__init__.py
@@ -1,3 +1,6 @@
 from .module import module
 
-from . import topgg
+from . import webhook
+from . import commands
+from . import data
+from . import settings
